@@ -2,14 +2,13 @@ apply plugin: 'com.android.library'
 apply plugin: 'com.github.dcendents.android-maven'
 apply plugin: 'com.jfrog.bintray'
 android {
-    compileSdkVersion 28
-//    buildToolsVersion '28.0.3'
+    compileSdkVersion 29
 
     defaultConfig {
-        minSdkVersion 11
-        targetSdkVersion 28
+        minSdkVersion 14
+        targetSdkVersion 29
         versionCode 1
-        versionName "1.0.1"
+        versionName "1.0.2"
         vectorDrawables.useSupportLibrary = true
     }
     buildTypes {
@@ -22,10 +21,10 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation('com.android.support:appcompat-v7:28.0.0') {
+    implementation('androidx.appcompat:appcompat:1.1.0') {
         exclude module: 'support-v4'
     }
-    implementation ('com.android.support:recyclerview-v7:28.0.0'){
+    implementation('androidx.recyclerview:recyclerview:1.1.0'){
         exclude module: 'support-v4'
     }
     implementation ('com.github.chrisbanes.photoview:library:1.2.4'){
@@ -41,7 +40,7 @@ def libName = "zImagePicker"
 //发布到组织名称名字，必须填写
 group = "com.zxn.imagepicker"
 // 版本号，下次更新是只需要更改版本号即可
-version = "1.0.1"
+version = "1.0.2"
 //上面配置后上传至JCenter后的编译路径是这样的： compile 'com.zxn.zatlib.zImagePicker:1.0.0'
 //生成源文件
 task sourcesJar(type: Jar) {
