@@ -1,6 +1,6 @@
 package com.zxn.imagepicker;
 
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 /**
  * 自定义一个Provider，以免和引入的项目的provider冲突
